+++ src/index.ts
@@ -1,5 +1,5 @@
 export { default, AppLovinMAX } from './AppLovinMAX';
-export { Privacy, ConsentFlowUserGeography, AppTrackingStatus } from './Privacy';
+export { Privacy, ConsentFlowUserGeography, AppTrackingStatus, CmpError } from './Privacy';
 export { TargetingData, AdContentRating, UserGender } from './TargetingData';
 export { InterstitialAd } from './InterstitialAd';
 export { RewardedAd } from './RewardedAd';